Ajout de fonctionnalités dans la notation

+++ AniSdoufBack/src/main/java/org/example/anisdoufback/dto/EpisodeResponse.java
@@ -18,6 +18,6 @@ public class EpisodeResponse implements Serializable {
     private Integer numero;
     private Collection<NoteEpisodeResponse> notesE;
     private Boolean estVu;
-    private Integer noteE;
+    private Double noteE;
     private Boolean estFavori;
 }
